moved cpp into fiducial_slam and killed python
