+++ fiducial_slam/include/fiducial_slam/helpers.h
