--- fiducial_slam/src/fiducial_slam/__init__.py
@@ -1,47 +1,0 @@
-
-import os, math
-
-from tf.transformations import euler_from_quaternion, quaternion_slerp, \
-                               translation_matrix, quaternion_matrix, \
-                               translation_from_matrix, quaternion_from_matrix, \
-                               quaternion_from_euler
-
-"""
-Create directory if not already there
-"""
-def mkdirnotex(filename):
-    dir=os.path.dirname(filename)
-    print "Directory", dir
-    if dir and not os.path.exists(dir):
-        os.makedirs(dir)
-
-"""
-Radians to degrees
-"""
-def rad2deg(rad):
-    return rad / math.pi * 180.0
-
-"""
-Degrees to radians
-"""
-def deg2rad(deg):
-    return float(deg) * math.pi / 180.0
-
-
-"""
-Weighted average of linear quantities
-"""
-def updateLinear(mean1, var1, mean2, var2):
-    newMean = (mean1 * var2 + mean2 * var1) / (var1 + var2)
-    newVar = 1.0 / (1.0/var1 + 1.0/var2)
-    return [newMean, newVar]
-
-
-"""
-Weighted average of quaternions
-"""
-def updateAngular(quat1, var1, quat2, var2):
-    newMean = quaternion_slerp(quat1, quat2, var1/(var1+var2))
-    newVar = 1.0 / (1.0/var1 + 1.0/var2)
-    return [newMean, newVar]
-
