--- fiducial_slam/src/fiducial_slam/fiducial.py
@@ -1,74 +1,0 @@
-#!/usr/bin/python
-
-"""
-Copyright (c) 2015, Ubiquity Robotics
-All rights reserved.
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-* Redistributions of source code must retain the above copyright notice, this
-  list of conditions and the following disclaimer.
-* Redistributions in binary form must reproduce the above copyright notice,
-  this list of conditions and the following disclaimer in the documentation
-  and/or other materials provided with the distribution.
-* Neither the name of ubiquity_motor nor the names of its
-  contributors may be used to endorse or promote products derived from
-  this software without specific prior written permission.
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""
-
-import rospy
-import numpy
-
-from tf.transformations import euler_from_quaternion, quaternion_slerp, \
-                               translation_matrix, quaternion_matrix, \
-                               translation_from_matrix, quaternion_from_matrix, \
-                               quaternion_from_euler
-from fiducial_slam import updateLinear, updateAngular
-
-
-"""
-Class to represent the pose and state of a single fiducial
-"""
-class Fiducial:
-    def __init__(self, id):
-        self.id = id
-        self.position = None
-        self.orientation = None
-        self.variance = None
-        self.observations = 0
-        self.links = []
-        self.publishedMarker = False
-        self.lastSeenTime = rospy.Time(0,0)
-
-    """
-    Return pose as a 4x4 matrix
-    """
-    def pose44(self):
-        return numpy.dot(translation_matrix(self.position),
-                         quaternion_matrix(self.orientation))
-
-    """
-    Update pose with a new observation
-    """
-    def update(self, newPosition, newOrientation, newVariance):
-        if self.observations == 0:
-            self.position = newPosition
-            self.orientation = numpy.array(newOrientation)
-            self.variance = newVariance
-            self.observations = 1
-            return
-        self.position, v1 = updateLinear(self.position, self.variance,
-                                         newPosition, newVariance)
-        self.orientation, v2 = updateAngular(self.orientation, self.variance,
-                                             newOrientation, newVariance)
-        self.variance = v1
-        self.observations += 1
