+++ fiducial_slam/test/111_initial_map.txt
