+++ fiducial_slam/test/aruco_images.bag
