+++ fiducial_slam/test/aruco_transforms.bag
