--- fiducial_slam/test/ceiling_test_2.xml
@@ -1,39 +1,0 @@
-<!--
-  Verify that playing a bag file of images with fiducials produces a map with the final
-  fiducial in a sane position
--->
-<launch>
-  <param name="/use_sim_time" value="true"/>
-
-  <node name="camera_pose" pkg="tf2_ros" type="static_transform_publisher" 
-        args="0.0 0.0 0.0 0.0 0.0 0.0 0.0 base_link raspicam" />
-
-  <node pkg="aruco_detect" name="aruco_detect" type="aruco_detect">
-    <param name="map_frame" value="odom" />
-    <param name="pose_frame" value="slam" />
-    <param name="image_transport" value="compressed" />
-    <param name="fiducial_len" value="0.145"/>
-    <remap from="/camera/compressed" to="/raspicam_node/image/compressed"/>
-    <remap from="/camera_info" to="/raspicam_node/camera_info"/>
-  </node>
-
-  <node type="fiducial_slam_node.py" pkg="fiducial_slam" name="fiducial_slam">
-    <param name="initial_map_file" value="$(find fiducial_slam)/test/initial_map2.txt" />
-    <param name="map_file" value="$(find fiducial_slam)/test/map2.txt" />
-    <param name="obs_file" value="$(find fiducial_slam)/test/obs.txt" />
-    <param name="trans_file" value="$(find fiducial_slam)/test/trans.txt" />
-    <param name="mapping_mode" value="true" />
-    <param name="camera_frame" value="base_link" />
-    <param name="odom_frame" value="" />
-    <param name="pose_frame" value="base_link" />
-  </node>
-
-  <node name="rosbag" pkg="rosbag" type="play" args="--clock -r 0.2 $(find fiducial_slam)/test/ceilingline.bag"/>
-
-  <test test-name="ceiling_map_test" pkg="fiducial_slam" type="map_test.py" name="map_test" time-limit="600.0">
-    <param name="map_file" value="$(find fiducial_slam)/test/map2.txt" />
-    <param name="min_lines" value="16" />
-    <param name="expect" value="16 0 9.7 0" />
-  </test>
-</launch>
-
