--- fiducial_slam/test/init_map.xml
@@ -1,32 +1,0 @@
-<!--
-  Verify that a map is created when a set of fiducual transforms is sent to
-  fiducial_slam
--->
-<launch>
-  <param name="/use_sim_time" value="true"/>
-
-  <node name="camera_pose" pkg="tf2_ros" type="static_transform_publisher" 
-        args="0.0 0.0 0.0 0.0 0.0 0.0 1.0 base_link raspicam" />
-
-  <node name="robot_pose" pkg="tf2_ros" type="static_transform_publisher" 
-        args="0.0 0.0 0.0 0.0 0.0 0.0 1.0 odom base_link" />
-
-  <node type="fiducial_slam_node.py" pkg="fiducial_slam" name="fiducial_slam">
-    <param name="map_file" value="$(find fiducial_slam)/test/map3.txt" />
-    <param name="obs_file" value="$(find fiducial_slam)/test/obs.txt" />
-    <param name="trans_file" value="$(find fiducial_slam)/test/trans.txt" />
-    <param name="mapping_mode" value="false" />
-    <param name="camera_frame" value="base_link" />
-    <param name="odom_frame" value="odom" />
-    <param name="pose_frame" value="base_link" />
-    <remap to="/fiducial_detect/fiducial_transforms" from="/fiducial_transforms"/>
-  </node>
-
-  <node name="rosbag" pkg="rosbag" type="play" args="--clock -l $(find fiducial_slam)/test/transforms.bag"/>
-
-  <test test-name="init_map_test" pkg="fiducial_slam" type="map_test.py" name="map_test3">
-    <param name="map_file" value="$(find fiducial_slam)/test/map3.txt" />
-    <param name="min_lines" value="6" />
-  </test>
-</launch>
-
