+++ fiducial_slam/test/init_map_aruco.xml
@@ -3,21 +3,21 @@
   fiducial_slam and no initial map is specified
 -->
 <launch>
-  <node type="fiducial_slam" pkg="fiducial_slam2" name="fiducial_slam">
+  <node type="fiducial_slam" pkg="fiducial_slam" name="fiducial_slam">
     <param name="initial_map_file" value="/dev/null" />
-    <param name="map_file" value="$(find fiducial_slam2)/test/map.txt" />
+    <param name="map_file" value="$(find fiducial_slam)/test/map.txt" />
     <param name="camera_frame" value="base_link" />
     <param name="odom_frame" value="" />
     <param name="pose_frame" value="base_link" />
   </node>
 
   <node name="rosbag" pkg="rosbag" type="play"
-    args="-l $(find fiducial_slam2)/test/aruco_transforms.bag">
+    args="-l $(find fiducial_slam)/test/aruco_transforms.bag">
   </node>
 
-  <test test-name="map_test3" pkg="fiducial_slam2" type="map_test.py"
+  <test test-name="map_test3" pkg="fiducial_slam" type="map_test.py"
      name="map_test">
-    <param name="map_file" value="$(find fiducial_slam2)/test/map.txt" />
+    <param name="map_file" value="$(find fiducial_slam)/test/map.txt" />
     <param name="min_lines" value="7" />
     <param name="expected_pose" value="0 0 0 0 0 0 1" />
   </test>
