--- fiducial_slam/test/initial_map.txt
@@ -1,1 +1,0 @@
-31 0 0 0 180 0 180 0 0
