--- fiducial_slam/test/initial_map2.txt
@@ -1,1 +1,0 @@
-1 0.0 0.0 0.0 180.0 -1.1761008961519854e-45 180.0 0.0 0
