--- fiducial_slam/test/localization_test.cpp
@@ -1,91 +1,0 @@
-#include <gtest/gtest.h>
-
-#include <ros/ros.h>
-#include <image_transport/image_transport.h>
-#include <sensor_msgs/CameraInfo.h>
-#include <opencv2/highgui/highgui.hpp>
-#include <cv_bridge/cv_bridge.h>
-
-#include <geometry_msgs/PoseWithCovarianceStamped.h>
-
-class LocalizationTest : public ::testing::Test {
-protected:
-  virtual void SetUp() { 
-    it = new image_transport::ImageTransport(nh);
-    image_pub = it->advertise("camera/image", 1);
-
-    CameraInfoPub = nh.advertise<sensor_msgs::CameraInfo>("camera_info", 5);
-
-    c_info.header.frame_id = "camera";
-
-    c_info.height = 960;
-    c_info.width = 1280;
-    c_info.distortion_model = "plumb_bob";
-
-    double data_D[] = {0.1349735087283542, -0.2335869827451621, 0.0006697030315075139, 0.004846737465872353, 0.0};
-    c_info.D = std::vector<double> (data_D, data_D + sizeof(data_D)/ sizeof(double));
-   
-    c_info.K = {1006.126285753055, 0.0, 655.8639244150409, 0.0, 1004.015433012594, 490.6140221242933, 0.0, 0.0, 1.0};
-    c_info.R = {1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0};
-    c_info.P = {1021.54345703125, 0.0, 661.9091982335958, 0.0, 0.0, 1025.251953125, 490.6380671707448, 0.0, 0.0, 0.0, 1.0, 0.0};
-
-    ros::NodeHandle nh_priv("~");
-    nh_priv.getParam("image_directory", image_directory);
-
-    pose_sub = nh.subscribe("/fiducial_pose", 1, &LocalizationTest::pose_callback, this);
-    got_pose = false;
-  }
-
-  virtual void TearDown() { delete it;}
-
-  void publish_image(std::string file) {
-    cv::Mat image = cv::imread(image_directory+file, CV_LOAD_IMAGE_COLOR);
-    cv::waitKey(30);
-    sensor_msgs::ImagePtr msg = cv_bridge::CvImage(std_msgs::Header(), "bgr8", image).toImageMsg();
-    image_pub.publish(msg);
-    c_info.header.stamp = ros::Time::now();
-    CameraInfoPub.publish(c_info);
-  }
-
-  void pose_callback(const geometry_msgs::PoseWithCovarianceStamped& p) {
-    got_pose = true;
-    pose = p;
-  }
-
-  ros::NodeHandle nh;
-
-  // Set up Publishing of static images
-  image_transport::ImageTransport* it;
-  image_transport::Publisher image_pub;
-
-  sensor_msgs::CameraInfo c_info;
-  ros::Publisher CameraInfoPub;
-
-  std::string image_directory;
-
-  // Set up subscribing
-  bool got_pose;
-  geometry_msgs::PoseWithCovarianceStamped pose;
-  ros::Subscriber pose_sub;
-};
-
-TEST_F(LocalizationTest, sparse_board) {
-  ros::Rate loop_rate(5);
-  while (nh.ok() && !got_pose) {
-    publish_image("sparse_board.png");
-    ros::spinOnce();
-    loop_rate.sleep();
-  }
-
-  ASSERT_NEAR(-0.164061, pose.pose.pose.position.x, 0.01);
-  ASSERT_NEAR(-0.181358, pose.pose.pose.position.y, 0.01);
-  ASSERT_NEAR(0, pose.pose.pose.position.z, 0.01);
-}
-
-int main(int argc, char** argv)
-{
-
-  testing::InitGoogleTest(&argc, argv);
-  ros::init(argc, argv, "LocalizationTest");
-  return RUN_ALL_TESTS();
-}
