--- fiducial_slam2/launch/fiducial_slam.launch
@@ -1,19 +1,0 @@
-<!--
-   Run the fiducial_slam node
--->
-<launch>
-  <arg name="map_frame" default="map"/>
-  <arg name="odom_frame" default="odom"/>
-  <arg name="base_frame" default="base_link"/>
-  <arg name="publish_tf" default="true"/>
-  <arg name="future" default="0.0"/>
-
-
-  <node type="fiducial_slam" pkg="fiducial_slam2" output="screen" 
-    name="fiducial_slam">
-    <param name="map_file" value="$(env HOME)/.ros/slam/map.txt" />
-    <param name="map_frame" value="$(arg map_frame)" />
-    <param name="odom_frame" value="$(arg odom_frame)" />
-    <param name="base_frame" value="$(arg base_frame)" />
-  </node>
-</launch>
