--- fiducial_slam2/package.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>fiducial_slam2</name>
-  <version>0.6.1</version>
-  <description>ROS node to build a 3D map of fiducials and estimate robot pose from fiducial transforms</description>
-
-  <maintainer email="jimv@mrjim.com">Jim Vaughan</maintainer>
-  <maintainer email="send2arohan@gmail.com">Rohan Agrawal</maintainer>
-
-  <license>BSD</license>
-
-  <author email="jimv@mrjim.com">Jim Vaughan</author>
-
-  <buildtool_depend>catkin</buildtool_depend>
-
-  <depend>roscpp</depend>
-  <depend>tf2_geometry_msgs</depend>
-  <depend>tf2_ros</depend>
-  <depend>tf2</depend>
-  <depend>visualization_msgs</depend>
-  <depend>image_transport</depend>
-  <depend>sensor_msgs</depend>
-  <depend>cv_bridge</depend>
-  <depend>opencv3</depend>
-  <depend>fiducial_msgs</depend>
-  <depend>dynamic_reconfigure</depend>
-
-</package>
