--- fiducial_slam2/test/610_initial_map.txt
@@ -1,1 +1,0 @@
-610 0 0 0 180 0 180 0 0
