--- fiducial_slam2/test/create_map.xml
@@ -1,31 +1,0 @@
-<!--
-  Verify that a map is created when a set of fiducual transforms is sent to
-  fiducial_slam
--->
-<launch>
-  <node name="camera_pose" pkg="tf2_ros" type="static_transform_publisher" 
-        args="0.0 0.0 0.0 0.0 0.0 0.0 1.0 base_link raspicam" />
-
-  <node type="fiducial_slam" pkg="fiducial_slam2" name="fiducial_slam">
-    <param name="initial_map_file" value="$(find fiducial_slam2)/test/610_initial_map.txt" />
-    <param name="map_file" value="$(find fiducial_slam2)/test/map.txt" />
-    <param name="odom_frame" value="" />
-    <param name="pose_frame" value="base_link" />
-  </node>
-
-  <node name="rosbag" pkg="rosbag" type="play" 
-     args="-l $(find fiducial_slam2)/test/board_test_transforms.bag">
-  </node>
-
-  <test test-name="map_test" pkg="fiducial_slam2" type="map_test.py" name="map_test1">
-    <param name="map_file" value="$(find fiducial_slam2)/test/map.txt" />
-    <param name="min_lines" value="6" />
-    <param name="expect" value="610 0.0 0.0 0.0 180.0 0.0 180.0, 
-                                611 0.0 -0.19 0.0 -178.4 -1.3 179.7,
-                                612 0.0 -0.37 0.0 -178.7 -0.3 179.9, 
-                                613 -0.17 0.0 -0.0 -179.0 1.5 -179.9,
-                                614 -0.18 -0.2 0.0 -178.3 1.0 179.5,
-                                615 -0.18 -0.4 0.0 -177.0 1.1 179.4"/>
-  </test>
-</launch>
-
