--- fiducial_slam2/test/map.txt
@@ -1,7 +1,0 @@
-100 0.005648 -1.107023 2.824314 145.729026 -21.882801 147.358662 0.000492 30 103 106 107 110 111 112
-103 -1.673025 0.391916 2.895842 179.124586 1.497907 163.875975 0.000000 10 100 106 107 110 111 112
-106 -0.120415 -0.035908 1.119976 -172.616040 20.598438 -0.087160 0.000228 30 100 103 107 110 111 112
-107 -0.112052 0.241676 1.174273 -172.323734 19.200012 0.846731 0.000160 30 100 103 106 110 111 112
-110 0.332209 -0.088042 0.942426 -155.786888 -3.267327 88.933802 0.000370 30 100 103 106 107 111 112
-111 -0.325674 -0.037940 1.196333 -171.647844 20.813304 0.111855 0.000086 30 100 103 106 107 110 112
-112 -0.317858 0.239751 1.247673 -172.708425 20.326869 0.086530 0.000011 30 100 103 106 107 110 111
